# Migrate some C code to Ruby