@eregon 2024-10-17
Nice cleanup!
I think it would be great to share it with the pure-Ruby backend, so we can optimize that part for both and deduplicate.
As we both noticed, the amount of work that goes into creating that State object seems way too much.