@headius 2024-10-17
Similar code in the JRuby extension could probably also be removed with this change.