@headius 2024-10-17
I took a quick stab at it but there's some complexities mixing the pure-Ruby and extension versions of State.

This PR only moves *some* of State into Ruby, so I attempted to only remove those parts of the JRuby State implementation. It failed because of a few things:

* The pure-Ruby version and the C-ext State and this new ext pure-Ruby State use instance variables to store all configuration. The JRuby extension uses Java fields to avoid the overhead of instance variables. The JRuby State would need to switch to using instance variables.
* The JRuby extension State does not defined `allow_nan=`. JRuby extension would need to add this (or these config ivars could all move into attrs). This probably also applies to some of the other config fields.

Mostly the JRuby ext would need to switch to instance vars for all accesses. That would be slower than using Java object fields, but it may not be important for `Generator::State` configuration.

The patch, such as it is: https://gist.github.com/headius/40b69293f3d34a9bfc1373675369dba8